requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
